+++ packages/client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Emit current finalized and descendants when subscribing to `blocks$`
+
 ## 1.21.2 - 2025-11-27
 
 ### Fixed